+++ packages/engine/debug/index.pug
@@ -19,4 +19,4 @@ html(lang="ja")
     p
         | &quot;
         a(href="http://wwawing.com") WWA Wing
-        | &quot; &copy; 2013-#{thisYear} WWA Wing Team
+        | &quot; &copy; 2013-#{page.thisYear} WWA Wing Team
